add decay for history edge weight

+++ inc/SharedEnv.h
@@ -40,6 +40,6 @@ public:
     SharedEnvironment(){}
 
     // vector<pair<int,int>> past_waitings;
-    vector<pair<double,int>> past_waitings;
+    vector<pair<double,double>> past_waitings;
     vector<int> accu_waitings;
 };
